Playing around further with Bootstrap.
I have a better layout built now, and have figured out how to use
bootstrap for a variety of effects. Unfortunately this doesn't seem to
work on the university webspace so I will need to use my own webspace to
test.

This may be linked to using the bootstrap CDN as opposed to downloading
it myself, but everything in my soul prefers the CDN option, so I will
persevere.

Next I need to fill the page with content as now the majority of the
layout is built.